[~] This commit should make no real change to the plugin itself, mainly commenting, rearranging and renaming code and fixing up some warnings

+++ Texture/Texture.h
@@ -27,10 +27,10 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 class CTexture;
 
 typedef struct {
-	unsigned short int	dwWidth;			// Describes the width of the real texture area. Use lPitch to move between successive lines
-	unsigned short int	dwHeight;			// Describes the height of the real texture area
-	unsigned short int	dwCreatedWidth;		// Describes the width of the created texture area. Use lPitch to move between successive lines
-	unsigned short int	dwCreatedHeight;	// Describes the height of the created texture area
+	unsigned int	dwWidth;			// Describes the width of the real texture area. Use lPitch to move between successive lines
+	unsigned int	dwHeight;			// Describes the height of the real texture area
+	unsigned int	dwCreatedWidth;		// Describes the width of the created texture area. Use lPitch to move between successive lines
+	unsigned int	dwCreatedHeight;	// Describes the height of the created texture area
 	LONG		lPitch;				// Specifies the number of bytes on each row (not necessarily bitdepth*width/8)
 	VOID		*lpSurface;			// Pointer to the top left pixel of the image
 } DrawInfo;
